refactor: indexes and file name
