+++ server/src/jobs/jobs.ts
@@ -35,9 +35,6 @@ import { relanceFormulaire } from "./lba_recruteur/formulaire/relanceFormulaire"
 import { importReferentielOpcoFromConstructys } from "./lba_recruteur/opco/constructys/constructysImporter"
 import { relanceOpco } from "./lba_recruteur/opco/relanceOpco"
 import { updateSiretInfosInError } from "./lba_recruteur/user/misc/updateSiretInfosInError"
-import updateGeoLocations from "./lbb/updateGeoLocations"
-import updateLbaCompanies from "./lbb/updateLbaCompanies"
-import updateOpcoCompanies from "./lbb/updateOpcoCompanies"
 import { createJobsCollectionForMetabase } from "./metabase/metabaseJobsCollection"
 import { createRoleManagement360 } from "./metabase/metabaseRoleManagement360"
 import { runGarbageCollector } from "./misc/runGarbageCollector"
@@ -60,6 +57,9 @@ import { premiumInviteOneShot } from "./rdv/premiumInviteOneShot"
 import { removeDuplicateEtablissements } from "./rdv/removeDuplicateEtablissements"
 import { syncEtablissementDates } from "./rdv/syncEtablissementDates"
 import { syncEtablissementsAndFormations } from "./rdv/syncEtablissementsAndFormations"
+import updateGeoLocations from "./recruteurLba/updateGeoLocations"
+import updateOpcoCompanies from "./recruteurLba/updateOpcoCompanies"
+import updateLbaCompanies from "./recruteurLba/updateRecruteurLba"
 import { importReferentielRome } from "./seed/referentielRome/referentielRome"
 import updateBrevoBlockedEmails from "./updateBrevoBlockedEmails/updateBrevoBlockedEmails"
 import { controlApplications } from "./verifications/controlApplications"
