+++ server/src/jobs/recruteurLba/updateGeoLocations.ts
@@ -14,7 +14,7 @@ import __dirname from "../../common/dirname"
 import { logMessage } from "../../common/utils/logMessage"
 import { notifyToSlack } from "../../common/utils/slackUtils"
 
-import { checkIfAlgoFileIsNew, downloadAlgoCompanyFile, readCompaniesFromJson, removePredictionFile } from "./lbaCompaniesUtils"
+import { checkIfAlgoFileIsNew, downloadAlgoCompanyFile, readCompaniesFromJson, removePredictionFile } from "./RecruteurLbaUtils"
 
 const currentDirname = __dirname(import.meta.url)
