+++ shared/models/index.ts
@@ -15,13 +15,13 @@ export * from "./etablissement.model"
 export * from "./formation.model"
 export * from "./geolocations.model"
 export * from "./job.model"
-export * from "./lbaCompany.model"
-export * from "./lbaCompanyLegacy.model"
 export * from "./lbaItem.model"
 export * from "./lbacError.model"
 export * from "./metiersdavenir.model"
 export * from "./optout.model"
 export * from "./recruiter.model"
+export * from "./recruteurLba.model"
+export * from "./recruteurLbaLegacy.model"
 export * from "./recruteurLbaUpdateEvent.model"
 export * from "./referentielOnisep.model"
 export * from "./referentielOpco.model"
@@ -30,7 +30,7 @@ export * from "./rome.model"
 export * from "./session.model"
 export * from "./siretDiffusibleStatus.model"
 export * from "./unsubscribeOF.model"
-export * from "./unsubscribedLbaCompany.model"
+export * from "./unsubscribedRecruteurLba.model"
 export * from "./user.model"
 export * from "./userWithAccount.model"
 export * from "./usersRecruteur.model"
