+++ shared/models/internalJob.model.ts
@@ -8,6 +8,12 @@ const ZInternalJobs = z.any()
 
 export default {
   zod: ZInternalJobs,
-  indexes: [],
+  indexes: [
+    [{ type: 1, scheduled_for: 1 }, {}],
+    [{ type: 1, status: 1, scheduled_for: 1 }, {}],
+    [{ type: 1, status: 1, worker_id: 1, started_at: 1 }, {}],
+    [{ type: 1, name: 1 }, {}],
+    [{ ended_at: 1 }, { expireAfterSeconds: 3600 * 24 * 90 }],
+  ],
   collectionName,
 } as const satisfies IModelDescriptor
